+++ module/Core/test/ShortUrl/Helper/ShortUrlTitleResolutionHelperTest.php
@@ -5,11 +5,14 @@ declare(strict_types=1);
 namespace ShlinkioTest\Shlink\Core\ShortUrl\Helper;
 
 use Exception;
+use Fig\Http\Message\RequestMethodInterface;
 use GuzzleHttp\ClientInterface;
+use GuzzleHttp\RequestOptions;
 use Laminas\Diactoros\Response;
 use Laminas\Diactoros\Response\JsonResponse;
 use Laminas\Diactoros\Stream;
 use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\MockObject\Builder\InvocationMocker;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Shlinkio\Shlink\Core\Options\UrlShortenerOptions;
@@ -18,6 +21,8 @@ use Shlinkio\Shlink\Core\ShortUrl\Model\ShortUrlCreation;
 
 class ShortUrlTitleResolutionHelperTest extends TestCase
 {
+    private const LONG_URL = 'http://foobar.com/12345/hello?foo=bar';
+
     private MockObject & ClientInterface $httpClient;
 
     protected function setUp(): void
@@ -28,7 +33,7 @@ class ShortUrlTitleResolutionHelperTest extends TestCase
     #[Test]
     public function dataIsReturnedAsIsWhenResolvingTitlesIsDisabled(): void
     {
-        $data = ShortUrlCreation::fromRawData(['longUrl' => 'http://foobar.com/12345/hello?foo=bar']);
+        $data = ShortUrlCreation::fromRawData(['longUrl' => self::LONG_URL]);
         $this->httpClient->expects($this->never())->method('request');
 
         $result = $this->helper()->processTitle($data);
@@ -40,7 +45,7 @@ class ShortUrlTitleResolutionHelperTest extends TestCase
     public function dataIsReturnedAsIsWhenItAlreadyHasTitle(): void
     {
         $data = ShortUrlCreation::fromRawData([
-            'longUrl' => 'http://foobar.com/12345/hello?foo=bar',
+            'longUrl' => self::LONG_URL,
             'title' => 'foo',
         ]);
         $this->httpClient->expects($this->never())->method('request');
@@ -53,10 +58,8 @@ class ShortUrlTitleResolutionHelperTest extends TestCase
     #[Test]
     public function dataIsReturnedAsIsWhenFetchingFails(): void
     {
-        $data = ShortUrlCreation::fromRawData([
-            'longUrl' => 'http://foobar.com/12345/hello?foo=bar',
-        ]);
-        $this->httpClient->expects($this->once())->method('request')->willThrowException(new Exception('Error'));
+        $data = ShortUrlCreation::fromRawData(['longUrl' => self::LONG_URL]);
+        $this->expectRequestToBeCalled()->willThrowException(new Exception('Error'));
 
         $result = $this->helper(autoResolveTitles: true)->processTitle($data);
 
@@ -66,10 +69,8 @@ class ShortUrlTitleResolutionHelperTest extends TestCase
     #[Test]
     public function dataIsReturnedAsIsWhenResponseIsNotHtml(): void
     {
-        $data = ShortUrlCreation::fromRawData([
-            'longUrl' => 'http://foobar.com/12345/hello?foo=bar',
-        ]);
-        $this->httpClient->expects($this->once())->method('request')->willReturn(new JsonResponse(['foo' => 'bar']));
+        $data = ShortUrlCreation::fromRawData(['longUrl' => self::LONG_URL]);
+        $this->expectRequestToBeCalled()->willReturn(new JsonResponse(['foo' => 'bar']));
 
         $result = $this->helper(autoResolveTitles: true)->processTitle($data);
 
@@ -79,10 +80,8 @@ class ShortUrlTitleResolutionHelperTest extends TestCase
     #[Test]
     public function dataIsReturnedAsIsWhenTitleCannotBeResolvedFromResponse(): void
     {
-        $data = ShortUrlCreation::fromRawData([
-            'longUrl' => 'http://foobar.com/12345/hello?foo=bar',
-        ]);
-        $this->httpClient->expects($this->once())->method('request')->willReturn($this->respWithoutTitle());
+        $data = ShortUrlCreation::fromRawData(['longUrl' => self::LONG_URL]);
+        $this->expectRequestToBeCalled()->willReturn($this->respWithoutTitle());
 
         $result = $this->helper(autoResolveTitles: true)->processTitle($data);
 
@@ -92,15 +91,29 @@ class ShortUrlTitleResolutionHelperTest extends TestCase
     #[Test]
     public function titleIsUpdatedWhenItCanBeResolvedFromResponse(): void
     {
-        $data = ShortUrlCreation::fromRawData([
-            'longUrl' => 'http://foobar.com/12345/hello?foo=bar',
-        ]);
-        $this->httpClient->expects($this->once())->method('request')->willReturn($this->respWithTitle());
+        $data = ShortUrlCreation::fromRawData(['longUrl' => self::LONG_URL]);
+        $this->expectRequestToBeCalled()->willReturn($this->respWithTitle());
 
         $result = $this->helper(autoResolveTitles: true)->processTitle($data);
 
         self::assertNotSame($data, $result);
         self::assertEquals('Resolved "title"', $result->title);
+    }
+
+    private function expectRequestToBeCalled(): InvocationMocker
+    {
+        return $this->httpClient->expects($this->once())->method('request')->with(
+            RequestMethodInterface::METHOD_GET,
+            self::LONG_URL,
+            [
+                RequestOptions::TIMEOUT => 3,
+                RequestOptions::CONNECT_TIMEOUT => 3,
+                RequestOptions::ALLOW_REDIRECTS => ['max' => ShortUrlTitleResolutionHelper::MAX_REDIRECTS],
+                RequestOptions::IDN_CONVERSION => true,
+                RequestOptions::HEADERS => ['User-Agent' => ShortUrlTitleResolutionHelper::CHROME_USER_AGENT],
+                RequestOptions::STREAM => true,
+            ],
+        );
     }
 
     private function respWithoutTitle(): Response